Add files via upload
I made a move function for the player, so now the player can jump and move around! But that’s only for player 1 for now because I’m gonna work on player 1 and then just do the same for player 2, so that’s fairly simple.
I also started creating the attack function for the player, I’m basically making a hitbox right now, but I kind of have an error with calling the function. I’m gonna fix that, but I made some progress so that’s good.