Add files via upload
So, I’m still stumped with the spritesheet. I’m trying to scale the image but it’s really weird considering the frame again isn’t a square. I’m going to try getting the spritesheet done by this weekend, but I’m also going to try using this other random spritesheet I found on the internet for the meanwhile and play around with it. Other than that, I also researched how to do audio on pygame.
I’m hoping to get the sprites done and whatnot by the weekend-ish? And then work on the video and voice lines/audio.