Add files via upload
Sooo, I got some of the sprite work done, so I’m definitely working on updating the code during the weekend. Here’s a little sneak peak. ;) (it's on the google doc journal LOL)

In any event, I cleaned up some of my code, and added music and an icon to the game. The music was actually a little iffy to do because I was having trouble running it at the appropriate time and changing the volume, but then I realized I just put it in the wrong place and there’s a function that can change the volume…The icon making was fairly easy though. 

Also, note, I used music from SCP 3000 (the roblox game), but I might make my own music later for the background. But I'm definitely going to do sound effects pertaining to voice lines (hopefully).