Add files via upload
Issue: sprite has a weird restriction with the screen, also sprite spazzes out when animation occurs for running.
The spritesheet is still a problem but I got the idle working for player one. The idle for player two has been uploaded and it “works” but the animation is weird in the sense that it’s not doing the animation fast enough or doing the animation in various places because the program doesn’t know where to blit it. I atried changing up some things to see how to fix this but alas, here we are… I’m honestly hoping to fix this spritesheet by this week or something. ^^; But considering I at least got the correct spritesheets to run when the correct action is happening is good.