Add files via upload
I GOT THE SPRITES TO WORK!! Both player one and player two are currently animating the idle animation. Woohoo! There’s still an issue with screen restriction like the image of the sprite can go off the screen but it’s not technically going off the screen because of the red rectangle. It’s hard to put into words, but you can see it and see why there’s this weird screen restriction now. I also have to fix the flipping for player two because it kind of messes up the attacking if I try fixing it. But that’s besides the point, I got the sprites to work!! Woohoo! I’m planning to fix the weird screen restriction and flipping tomorrow, and then work on attacking animations and stuff.