Add files via upload
I’m still having issues with the sprites, I was analyzing the entirety of my code today, and I understand what the problem is but I’m not sure how to fix it exactly. I made some minor tweaks to test things out, but I didn’t come to a conclusion today. I’m going to try and get this fixed by tomorrow and so this weekend I can fully put all sprites in and add the intro video. If time permits, I’ll add sound effects and voice lines but that’s not my concern right now, the functionality of my program is.