Add files via upload
Added the running animation for the pink girl, but the atk animation for the blue girl is still an issue. The game can work now after the blue girl attacks but the animation never plays and she does attack for a certain amount of time but again the animation never plays. So, I’m a little stumped. I could give up with the attack animation and finish the game but I’m determined to fix it because if I fix it now, I can set up the pink girl animation easily and I have a clear cut plan of what to do for the intro/ending of the game.