Add files via upload
Not sure if this will update my files correctly but we'll see. >>>

Holy moly. I finally fixed my code. The animation for the attack actually works now and I even applied an attack cooldown (though my brother told me to make it less LOL). But man oh man, this is something I was stuck on for such a long time and I finally fixed it. Now all I have to do is the death animation followed by like the win screen and intro stuff. I’m really proud that I was able to fix that though because my game is actually functional now and doesn’t crash. (Somehow when I was working with the attack cooldown, I made an infinite loop and it kept crashing my program–my mac wasn’t the problem it was me LOL (joking))