Add files via upload
IT MAY 12:27 AM ON JUNE 11 RIGHT NOW BUT IVE JUST HAD SO MUCH FUN AND PRIDE FIGURING THIS OUT – I FINALLY GOT THE VIDEO TO PLAY FOR MY PROGRAM AHHH, the animation was made by my friend and IM JUST SO HAPPY. I will not lie most of the video code was from just researching a lot about different methods, I ended up going with importing cv2 because when I tried importing pyvidplayer 2 or moviepy, it kept giving me errors abut ffmpeg, which I’m not too fond of, but it was weird because I installed ffmpeg? I’m not entirely sure what that error was but I found that cv2 is just easier to work with in this case. I feel like this is so silly but I got this to work MUCH more easier than me handling the animations, but it was because I had this idea for a long time (of playing a video) so. I’m just sososo glad I got this to work. Oh yeah, I also added a death animation, scores, an intro timer (so there can be rounds and points now) and text. I just am so proud. Ahhhh!!